MAINT: linalg: move _ensure_dtype_cdsz to lapack.py
